Added more UI changes. Prediction graph mostly works. Still need to get dynamic resizing to work.
Added a 'boost' option to the loader. This allows for the pause to be
ignored for 'boost' seconds. The boost is not walltime, but rather is
relative to the timestamps being read from the input file. Thus, if 300
seconds is specified, then the first 5 minutes of data will be streamed
in as fast as possible after which the pause will be honored.